DEENV-147: Add file browser option for selecting a command in Sampling Settings (#157)
A new button is added to the Sampling Settings UI beneath the command input, labelled "Browse". Clicking this opens a file picker, to simplify selecting the command to sample.

To access the VS Code API, the call to open the file picker is on the extension backend. To implement this, two new message types were added: one from the webview to request the file picker to be opened, and one to the webview to update the view with the user's selection.

Also, the jest setup file and transformers were updated to support testing React components that use VS Code UI Toolkit components.