DEENV-150: Add validation function for RecordOptions (#169)
The new function will be used to validate that the user has set all of the required fields in sampling settings before running wperf record.

Also, functions and constants associated with the record options are moved into record-options to live with the type definition. This is because they will be used from both the webview and the extension backend, as well as it making a bit more sense logically.