DEENV-190: Add event table to Sampling Settings (#207)
Add the event table from the design to the form. It does not add the "edit row" (controls for adding events), that will follow in another commit.

The event description is shown as a tooltip.

There are action buttons that appear on hover for each row:

Delete is hooked up to remove selected events.
Edit currently does nothing.