DEENV-190: Store recent events when recording (#210)
We need recent events to complete the design. This PR adds a store that keeps track of the 5 most recently used events, every time the user hits record.

I pulled out a generic MementoStore class based on our old RecordOptionsStore class.

The recent events are sent to the Sampling Settings frontend and stored in the React state. They will be used in a follow up PR.