DEENV-234: Add function to run wperf test and parse the result (#254)
wperf test gives us some useful info, such as the default sampling frequency and the maximum number events we can sample. This PR adds a function for running wperf test and parsing the JSON output. We already had a function for getting text output used in the Check Driver Output command. This doesn't change that.